+++ Dockerfile.cuda
@@ -48,6 +48,7 @@ COPY pyproject.toml /app/pyproject.toml
 COPY uv.lock /app/uv.lock
 COPY README.md /app/README.md
 COPY src/ /app/src/
+COPY packages/ /app/packages/
 COPY configs /app/configs
 COPY examples /app/examples
 COPY benchmarks/scripts /app/benchmarks/scripts
