+++ packages/prime-rl-configs/pyproject.toml
@@ -1,0 +1,23 @@
+[project]
+name = "prime-rl-configs"
+version = "0.4.0"
+description = "Slim config schema for prime-rl (no GPU/ML deps)."
+readme = "README.md"
+requires-python = "~=3.12.0"
+dependencies = [
+    "loguru>=0.7.3",
+    "pydantic>=1.10.13",
+    "pydantic-config",
+    "tomli>=2.2.1",
+    "tomli-w>=1.2.0",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/prime_rl"]
+
+[tool.hatch.metadata]
+allow-direct-references = true
