+++ packages/prime-rl-configs/src/prime_rl/configs/orchestrator.py
@@ -2,8 +2,6 @@ import math
 from pathlib import Path
 from typing import Annotated, Any, Literal, TypeAlias
 
-from pydantic import AliasChoices, BaseModel, ConfigDict, Field, model_validator
-
 from prime_rl.configs.shared import (
     BaseModelConfig,
     ClientConfig,
@@ -18,6 +16,7 @@ from prime_rl.configs.shared import (
 from prime_rl.configs.trainer import TokenizerConfig
 from prime_rl.utils.config import BaseConfig
 from prime_rl.utils.logger import get_logger
+from pydantic import AliasChoices, BaseModel, ConfigDict, Field, model_validator
 
 
 class OptimizerConfig(BaseConfig):
