+++ packages/prime-rl-configs/src/prime_rl/configs/shared.py
@@ -2,9 +2,8 @@ import os
 from pathlib import Path
 from typing import Annotated, Literal, TypeAlias
 
-from pydantic import BaseModel, Field, model_validator
-
 from prime_rl.utils.config import BaseConfig
+from pydantic import BaseModel, Field, model_validator
 
 
 class SlurmConfig(BaseConfig):
