+++ pyproject.toml
@@ -5,16 +5,12 @@ description = "Async RL Training at Scale"
 readme = "README.md"
 requires-python = "~=3.12.0"
 dependencies = [
+    "prime-rl-configs",
     "beartype>=0.21.0",
     "datasets>=4.0.0",
     "jaxtyping>=0.3.2",
     "liger-kernel>=0.5.10",
-    "loguru>=0.7.3",
     "pyarrow>=21.0.0",
-    "pydantic>=1.10.13",
-    "pydantic-config",
-    "tomli>=2.2.1",
-    "tomli-w>=1.2.0",
     "numpy>=2.2.6",
     "torch>=2.9.0",
     "torchdata>=0.11.0",
@@ -108,6 +104,9 @@ dev = [
 ]
 
 
+[tool.uv.workspace]
+members = ["packages/*"]
+
 [tool.uv]
 exclude-newer = "7 days"
 preview = true
@@ -158,6 +157,7 @@ deep-gemm = false
 nixl-cu12 = false
 
 [tool.uv.sources]
+prime-rl-configs = { workspace = true }
 torch = { index = "pytorch-cu128" }
 verifiers = { git = "https://github.com/PrimeIntellect-ai/verifiers.git", rev = "3b77145" }
 torchtitan = { git = "https://github.com/pytorch/torchtitan", rev = "a1fdd7e" }
@@ -227,6 +227,9 @@ markers = [
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/prime_rl"]
 
 [tool.hatch.metadata]
 allow-direct-references = true
