--- src/prime_rl/__init__.py
@@ -1,1 +1,0 @@
-import prime_rl._compat  # noqa: F401 — must run before ring_flash_attn is imported
